% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{ff_f}
\alias{ff_f}
\title{functional form f}
\usage{
ff_f(c, mu, sigma)
}
\arguments{
\item{c}{maximum value of function}

\item{mu}{optimal temperature}

\item{sigma}{temperature tolerance}
}
\description{
Return a function that takes a single argument \code{w} (current temperature).
}
