% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{ff_g}
\alias{ff_g}
\title{functional form g}
\usage{
ff_g(c, mu, sigma)
}
\arguments{
\item{c}{minimum value of function}

\item{mu}{optimal temperature}

\item{sigma}{temperature tolerance}
}
\description{
Return a function that takes a single argument \code{w} (current temperature).
}
