% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions_generator.R
\name{generate_egg_advancement_transitions}
\alias{generate_egg_advancement_transitions}
\title{Generate egg advancement transitions}
\usage{
generate_egg_advancement_transitions(cube, u, nE)
}
\arguments{
\item{cube}{an inheritance cube (see \link[MGDrivE:cubeHomingDrive]{MGDrivE::cubeHomingDrive} for an example)}

\item{u}{character vector of places}

\item{nE}{number of egg stages (shape parameter)}
}
\description{
Generates egg advancement transitions for a single node
}
