% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions_generator.R
\name{generate_egg_mortality_transitions}
\alias{generate_egg_mortality_transitions}
\title{Generate egg mortality transitions}
\usage{
generate_egg_mortality_transitions(cube, u, nE)
}
\arguments{
\item{cube}{an inheritance cube (see \link[MGDrivE:cubeHomingDrive]{MGDrivE::cubeHomingDrive} for an example)}

\item{u}{character vector of places}

\item{nE}{number of egg stages (shape parameter)}
}
\description{
Generates egg mortality transitions for a single node
}
