% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions_generator.R
\name{generate_oviposition_transitions}
\alias{generate_oviposition_transitions}
\title{Generate oviposition transitions}
\usage{
generate_oviposition_transitions(cube, u)
}
\arguments{
\item{cube}{an inheritance cube (see \link[MGDrivE:cubeHomingDrive]{MGDrivE::cubeHomingDrive} for an example)}

\item{u}{character vector of places}
}
\description{
Generates oviposition transitions for a single node
}
