% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{make_beta}
\alias{make_beta}
\title{female fecundity rate}
\usage{
make_beta(c, mu, sigma, s)
}
\arguments{
\item{c}{maximum value of function}

\item{mu}{optimal temperature}

\item{sigma}{temperature tolerance}

\item{s}{genotype specific multiplier of fecundity}
}
\description{
This uses functional form \code{f} but also the genotype specific multiplier \code{s}
}
