% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{make_lifecycle_parameters}
\alias{make_lifecycle_parameters}
\title{Make d. suzukii life history parameters}
\usage{
make_lifecycle_parameters(
  c_beta = 1.6,
  mu_beta = 21.9,
  sigma_beta = 6.1,
  c_E = 1.3,
  c_L = 5.1,
  c_P = 4.6,
  mu_E = 25.3,
  mu_L = 25.3,
  mu_P = 25.3,
  sigma_E = 11.1,
  sigma_L = 11.1,
  sigma_P = 11.1,
  k_E = 15,
  k_L = 25,
  k_P = 25,
  c_A = 22,
  mu_A = 5,
  sigma_A = 25.9,
  k_A = 4,
  c_delta = 0.014,
  mu_delta = 17.7,
  sigma_delta = 7.8
)
}
\arguments{
\item{c_beta}{max daily fecundity}

\item{mu_beta}{optimal temperature for fecundity}

\item{sigma_beta}{temperature tolerance for fecundity}

\item{c_E}{min duration of egg stage}

\item{c_L}{min duration of larvae stage}

\item{c_P}{min duration of pupae stage}

\item{mu_E}{optimal temperature for egg stage}

\item{mu_L}{optimal temperature for larvae stage}

\item{mu_P}{optimal temperature for pupae stage}

\item{sigma_E}{temperature tolerance for egg stage}

\item{sigma_L}{temperature tolerance for larvae stage}

\item{sigma_P}{temperature tolerance for pupae stage}

\item{k_E}{number of egg substages}

\item{k_L}{number of larvae substages}

\item{k_P}{number of pupae substages}

\item{c_A}{max duration of adult stage}

\item{mu_A}{optimal temperature of adult stage}

\item{sigma_A}{temperature tolerance of adult stage}

\item{k_A}{number of adult substages (not used)}

\item{c_delta}{min juvenile mortality}

\item{mu_delta}{optimal temperature for juvenile mortality}

\item{sigma_delta}{temperature tolerance for juvenile mortality}
}
\description{
Make d. suzukii life history parameters
}
