% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPN-hazard.R
\name{spn_hazards}
\alias{spn_hazards}
\title{Make hazards}
\usage{
spn_hazards(spn_P, spn_T, cube, par, exact = TRUE, tol = 1e-12, pbar = TRUE)
}
\arguments{
\item{spn_P}{set of places}

\item{spn_T}{set of transitions}

\item{cube}{inheritance cube}

\item{par}{parameters}

\item{exact}{return exact or approxiamte hazards}

\item{tol}{numeric tolerance to clip approximate hazards to zero}

\item{pbar}{progress bar when making hazards}
}
\description{
Make hazards
}
