% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPN-P.R
\name{spn_P}
\alias{spn_P}
\title{Build the set of places of the SPN}
\usage{
spn_P(parameters, cube)
}
\arguments{
\item{parameters}{lifecycle parameters (named \link{list})}

\item{cube}{an inheritance cube (named \link{list})}
}
\description{
Build the set of places of the SPN
}
