% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PN-lifecycle-node-P.R
\name{spn_P_lifecycle_node}
\alias{spn_P_lifecycle_node}
\title{Make Places (P) For a Node (D. Suzukii only)}
\usage{
spn_P_lifecycle_node(params, cube)
}
\arguments{
\item{params}{a named list of parameters (see details)}

\item{cube}{an inheritance cube from the \code{MGDrivE} package (e.g. \code{\link[MGDrivE]{cubeMendelian}})}
}
\value{
a list with two elements: \code{ix} contains labeled indices of the places
by life stage, \code{u} is the character vector of places (P)
}
\description{
This function makes the set of places (P) for a SPN. It is used alone if our model
is a single-node metapopulation for d. suzukii dynamics only.
}
\details{
The \code{params} argument supplies all of the ecological parameters necessary
to calculate equilibrium values. This function requires the \code{nE},
\code{nL}, \code{nA}, and \code{nP} parameters to be specified.
}
