% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPN-Matrix.R
\name{spn_Pre}
\alias{spn_Pre}
\title{Make Pre Matrix For a Petri Net}
\usage{
spn_Pre(spn_P, spn_T)
}
\arguments{
\item{spn_P}{set of places (P) (see details)}

\item{spn_T}{set of transitions (T) (see details)}
}
\value{
a matrix of type \code{\link[Matrix]{dgCMatrix-class}}
}
\description{
Generate the Pre (|v| by |u|) matrix for the SPN. This gives the edges from P
to T (input arcs) in the bipartite network.
}
