% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPN-Matrix.R
\name{spn_S}
\alias{spn_S}
\title{Make stoichiometry Matrix For a Petri Net}
\usage{
spn_S(spn_P, spn_T)
}
\arguments{
\item{spn_P}{set of places (P) (see details)}

\item{spn_T}{set of transitions (T) (see details)}
}
\description{
Generate the stoichiometry (|u| by |v|) matrix for the SPN.
Each column gives the net effect of that transition firing upon the state
space of the model. Internally, this creates a Pre (\code{\link{spn_Pre}}) and
Post (\code{\link{spn_Post}}) matrix, and then calculates the final stoichiometry.
}
