% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPN-T.R
\name{spn_T}
\alias{spn_T}
\title{Build the set of transitions of the SPN}
\usage{
spn_T(u, parameters, cube)
}
\arguments{
\item{u}{character vector of places}

\item{parameters}{lifecycle parameters (named \link{list})}

\item{cube}{an inheritance cube (named \link{list})}
}
\description{
Build the set of transitions of the SPN
}
