% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{suzukii_summarize_females}
\alias{suzukii_summarize_females}
\title{Return females from simulation}
\usage{
suzukii_summarize_females(out, spn_P)
}
\arguments{
\item{out}{an \link{array}}

\item{spn_P}{set of places}
}
\description{
Collapse (sum) over male mate genotyp
}
